W chuj zmian

+++ src/Random.java
@@ -3,7 +3,7 @@ import java.util.concurrent.ThreadLocalRandom;
 public class Random {
 
     public static double RInt(double min, double max){
-        double random = ThreadLocalRandom.current().nextDouble(min, max + 1);
+        double random = (int)ThreadLocalRandom.current().nextDouble(min, max + 1);
 
         return random;
     }
